refactor: TodoItemWrapper 컴포넌트 분리
- 구현 상세 제거
